requirements.txt
torch>=1.1
torchvision~=0.14.1
numpy~=1.21.4
tqdm~=4.64.1
tensorboard>=1.14

pandas~=1.3.4
Pillow~=8.4.0
pycocotools~=2.0.6